+++ CHANGELOG.md
@@ -12,8 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `tnt_net_thread_<index>_`
-Metrics for reporting net statistics per thread.
+- label `thread` for per thread reporting net statistics metrics.
 
 ## [0.14.0] - 2022-06-28
 ### Fixed